scribe/worker/requirements.txt
grpcio==1.72.0
grpcio-tools==1.72.0
protobuf>=5.29.0
faster-whisper==1.1.1
numpy>=1.26.0
torch>=2.4.0
sentence-transformers>=3.4.0
redis>=5.2.0
soundfile>=0.13.0
librosa>=0.10.0
